requirements.txt
# AI Agent Service Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.7.8
redis==5.0.1
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security and Monitoring
cryptography==41.0.8
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0